refactor: move room card runtime class
